pyproject.toml
[tool.poetry]
name = "google-kaggle"
version = "0.1.0"
description = ""
authors = ["Yury <yury@iterative.ai>"]
packages = [
    { include = "src" },
]
[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.12.0"
dvc = {extras = ["s3"], version = "^2.16.0"}
torchvision = "^0.13.0"
PyYAML = "^6.0"
matplotlib = "^3.5.2"
transformers = "^4.21.1"
dvclive = "^0.10.0"
split-folders = "^0.5.1"
numba = "0.56"  # this is needed to install umap-learn. without it umap installs earlier version of this package which is broken 
umap-learn = "^0.5.3"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
